Move cache tests to nightly
